extracted weaver code to multiple files
